[Refactor](Workspace): Moved configs to subfolder.
