--- lua/config/toggleterm.lua
@@ -1,1 +1,0 @@
-return {}
