+++ lua/ds_omega/layers/Workspace/incline.lua
@@ -1,3 +1,3 @@
 local simple_plugin_setup = require('ds_omega.utils').simple_plugin_setup
 
-simple_plugin_setup('incline', 'incline')
+simple_plugin_setup('incline', 'Workspace')
