+++ lua/ds_omega/layers/Workspace/jabs.lua
@@ -1,3 +1,3 @@
 local simple_plugin_setup = require('ds_omega.utils').simple_plugin_setup
 
-simple_plugin_setup('jabs', 'jabs')
+simple_plugin_setup('jabs', 'Workspace')
