+++ lua/ds_omega/layers/Workspace/lualine.lua
@@ -1,6 +1,6 @@
 local simple_plugin_setup = require('ds_omega.utils').simple_plugin_setup
 
-simple_plugin_setup('lualine', 'lualine')
+simple_plugin_setup('lualine', 'Workspace')
 
 -- local prequire = require('utils').prequire;
