+++ lua/ds_omega/layers/Workspace/true_zen.lua
@@ -1,3 +1,3 @@
 local simple_plugin_setup = require('ds_omega.utils').simple_plugin_setup
 
-simple_plugin_setup('true-zen', 'true_zen')
+simple_plugin_setup('true-zen', 'Workspace.true_zen')
