+++ lua/ds_omega/layers/Workspace/twilight.lua
@@ -1,3 +1,3 @@
 local simple_plugin_setup = require('ds_omega.utils').simple_plugin_setup
 
-simple_plugin_setup('twilight', 'twilight')
+simple_plugin_setup('twilight', 'Workspace')
